pyproject.toml
[tool.poetry]
name = "lihua-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Li Hua <lihua@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.8"
python-telegram-bot = {extras = ["job-queue", "rate-limiter", "webhooks"], version = "^21.2"}
cryptography = "^42.0.7"

[tool.isort]
#profile = "hug"
src_paths = ["lihua_telegram_bot", "tests"]

[tool.pylint.main]
jobs = 0
py-version = "3.12"
recursive = true
suggestion-mode = true

[tool.pylint.basic]
good-names = ["i", "j", "k", "ex", "Run", "_"]

[tool.pylint.classes]
check-protected-access-in-special-methods = true

[tool.pylint.format]
expected-line-ending-format = "LF"
ignore-long-lines = "^\\s*(# )?<?https?://\\S+>?$"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.pylint.logging]
logging-format-style = "new"

[tool.pylint."messages control"]
disable = ["missing-module-docstring", "missing-function-docstring"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
